+++ locallinker.h
@@ -8,8 +8,8 @@ namespace rrl {
     public:
         using Linker::Linker;
 
-        virtual uint64_t resolve_symbol(Library &library, std::string const &symbol_library, std::string const &symbol_name) override;
-        virtual void add_export(Library &library, std::string const &symbol, uint64_t address) override;
+        virtual uintptr_t resolve_symbol(Library &library, std::string const &symbol_library, std::string const &symbol_name) override;
+        virtual void add_export(Library &library, std::string const &symbol, uintptr_t address) override;
     };
 
 }
